Reworked patchMessageById to pass all tests

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long>
+public interface AccountRepository extends JpaRepository<Account, Integer>
 {
     /**
      * Retrieve an account by its username and password
